refactor: eliminate redundant parameter type mapping
Fixed the root cause of parameter type override issues by removing
redundant type mapping in MapQueryColumns. Parameters are already
properly mapped by the query analyzer using database introspection,
so re-mapping them caused the int override for limit/offset to be lost.

Changes:
- Renamed MapQueryColumns to MapResultColumns to clarify it only maps
  SELECT result columns, not parameters
- Removed redundant parameter type mapping (lines 260-274)
- Fixed codegen to use range by index to avoid value copies
- Updated service layer to use int instead of int32 for limit parameters
- All integration tests now pass with proper int type for limit/offset

The generated code now correctly uses int for limit/offset parameters
throughout the entire stack: generated queries, repository layer, and
service layer.

Resolves #43

+++ example-app/repository/post_repository.go
@@ -143,7 +143,7 @@ func (r *PostRepository) PublishPost(ctx context.Context, postID uuid.UUID) erro
 // GetFeaturedPosts returns posts marked as featured with custom business logic
 func (r *PostRepository) GetFeaturedPosts(ctx context.Context, limit int) ([]domain.PostSummary, error) {
 	// Use the generated GetPublishedPosts as a base, then filter
-	posts, err := r.GetPublishedPosts(ctx, int32(limit*2)) // Get more to filter
+	posts, err := r.GetPublishedPosts(ctx, limit*2) // Get more to filter
 	if err != nil {
 		return nil, fmt.Errorf("failed to get published posts for featured filtering: %w", err)
 	}
@@ -163,7 +163,7 @@ func (r *PostRepository) GetFeaturedPosts(ctx context.Context, limit int) ([]dom
 // GetTrendingPosts returns posts with high engagement with custom business logic
 func (r *PostRepository) GetTrendingPosts(ctx context.Context, limit int) ([]domain.PostSummary, error) {
 	// Use the generated GetPublishedPosts as a base, then apply trending logic
-	posts, err := r.GetPublishedPosts(ctx, int32(limit*3))
+	posts, err := r.GetPublishedPosts(ctx, limit*3)
 	if err != nil {
 		return nil, fmt.Errorf("failed to get published posts: %w", err)
 	}
@@ -212,7 +212,7 @@ func (r *PostRepository) GetPostSummary(ctx context.Context, id uuid.UUID) (stri
 func (r *PostRepository) GetPostsByTag(ctx context.Context, tagName string, limit int) ([]domain.PostSummary, error) {
 	// In a real implementation, this would use a proper SQL query
 	// For demo purposes, we'll use the generated method and filter
-	posts, err := r.GetPublishedPosts(ctx, int32(limit*3))
+	posts, err := r.GetPublishedPosts(ctx, limit*3)
 	if err != nil {
 		return nil, fmt.Errorf("failed to get posts for tag filtering: %w", err)
 	}
